requirements.txt
matplotlib==3.8.2
numpy==1.26.2
pandas==2.1.3
plotly==5.18.0
seaborn==0.13.0
snowflake_connector_python==3.5.0
streamlit==1.29.0
